services/load-generator/pyproject.toml
[project]
name = "load-generator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "ddtrace>=2.9",
    "confluent-kafka>=2.4",
    "httpx>=0.27",
    "pydantic>=2.7",
    "pyyaml>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
